fix: download conditions
