+++ cmd/kubectl-testkube/commands/testsuites/common.go
@@ -408,8 +408,10 @@ func DownloadArtifacts(id, dir, format string, masks []string, client apiclientv
 	for _, execution := range testSuiteExecution.ExecuteStepResults {
 		if execution.Step != nil {
 			for _, step := range execution.Execute {
-				if step.Execution != nil {
-					tests.DownloadArtifacts(step.Execution.Id, filepath.Join(dir, step.Execution.Id), format, masks, client)
+				if step.Execution != nil && step.Step != nil && step.Step.Test != "" {
+					if step.Execution.IsPassed() || step.Execution.IsFailed() {
+						tests.DownloadArtifacts(step.Execution.Id, filepath.Join(dir, step.Execution.Id), format, masks, client)
+					}
 				}
 			}
 		}
